fix: rename `types` to `operate_types`
