+++ operate/ledger/ethereum.py
@@ -24,7 +24,7 @@ import typing as t
 from aea_ledger_ethereum import EthereumApi, EthereumCrypto
 
 from operate.ledger.base import LedgerHelper
-from operate.types import LedgerType
+from operate.operate_types import LedgerType
 
 
 class Ethereum(LedgerHelper):
